microservices/audit-service/requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
dapr-ext-fastapi==1.16.0
pydantic==2.9.2
python-dotenv==1.0.1
sqlmodel==0.0.22
psycopg2-binary==2.9.10